case uid_t to unsigned long in printf format
git-svn-id: svn://svn.h5l.se/heimdal/trunk/heimdal@15246 ec53bebd-3082-4978-b11e-865c3cabbd6b

+++ appl/kf/kfd.c
@@ -235,7 +235,8 @@ proto (int sock, const char *service)
     if (tk_file.length != 1)
 	snprintf (ccname, sizeof(ccname), "%s", (char *)(tk_file.data));
     else
-	snprintf (ccname, sizeof(ccname), "FILE:/tmp/krb5cc_%u",pwd->pw_uid);
+	snprintf (ccname, sizeof(ccname), "FILE:/tmp/krb5cc_%lu",
+		  (unsigned long)pwd->pw_uid);
 
     status = krb5_cc_resolve (context, ccname, &ccache);
     if (status) {
